feat(ProjectCard): Enable raw hours editing for assigned or logged editors
This change introduces a new memoized `canEditRaw` value to the `ProjectCard` component. This value determines if the "Remaining RAW" input field should be editable. The input is now editable if the logged-in editor is assigned to the project OR if the logged-in editor has logged any hours for that project. This provides more flexibility for editors to update RAW hours even if they are not the primary assignee.